1. Fixed bugs from last meeting
2. Fixed image and nav-bar size and changed link color to white

3. Added set up and linked social media icons. Google plus is set up with my email and contact info so let me know if we need to update this later. Linked in can not be setup with a gmail account so I will leave it off until we have our own email set up.

4. Found a really cool contact us template and added it to the site.Please provide feedback. There is also a form to email us directly from the app that I have left out for now but would like to add in later.

5. Become a vendor submission now goes to the grooming standard gmail account.

6. Admin portal is now known as vendor survey and is now located at grooming-standard-dev.appspot.com/#/vendor survey.

This change is because vendors can enter survey information meaning it is no longer an admin only page. We can still access it at the same spot or choose to share the link. There is no way to navigate to the page without the link or knowing the address mentioned above.I will still add validation so that we or the vendors must log in later. Please let me know if you have questions on any of this.

7. Updated text on rankings page. Please provide feedback on this and let me know what to change. Also added some styling to the table and added user answers as reference.

8. Added question 12 to rankings page and vendor survey.

9. Added button to clear selection on rankings page.

10. Fixed rotating images

11.